Change package of :detekt-report-* to dev.detekt.report.*
