+++ detekt-core/src/test/kotlin/io/gitlab/arturbosch/detekt/core/reporting/OutputReportsSpec.kt
@@ -2,9 +2,9 @@ package io.gitlab.arturbosch.detekt.core.reporting
 
 import dev.detekt.api.Detektion
 import dev.detekt.api.OutputReport
-import io.github.detekt.report.html.HtmlOutputReport
-import io.github.detekt.report.md.MdOutputReport
-import io.github.detekt.report.xml.XmlOutputReport
+import dev.detekt.report.html.HtmlOutputReport
+import dev.detekt.report.md.MdOutputReport
+import dev.detekt.report.xml.XmlOutputReport
 import io.github.detekt.test.utils.resourceAsPath
 import io.github.detekt.tooling.dsl.ReportsSpecBuilder
 import io.gitlab.arturbosch.detekt.core.createNullLoggingSpec
