+++ detekt-report-html/src/main/resources/META-INF/services/dev.detekt.api.OutputReport
@@ -1,1 +1,1 @@
-io.github.detekt.report.html.HtmlOutputReport
+dev.detekt.report.html.HtmlOutputReport
