+++ detekt-report-md/src/main/resources/META-INF/services/dev.detekt.api.OutputReport
@@ -1,1 +1,1 @@
-io.github.detekt.report.md.MdOutputReport
+dev.detekt.report.md.MdOutputReport
