+++ detekt-report-sarif/src/main/resources/META-INF/services/dev.detekt.api.OutputReport
@@ -1,1 +1,1 @@
-io.github.detekt.report.sarif.SarifOutputReport
+dev.detekt.report.sarif.SarifOutputReport
