+++ detekt-report-xml/src/main/resources/META-INF/services/dev.detekt.api.OutputReport
@@ -1,1 +1,1 @@
-io.github.detekt.report.xml.XmlOutputReport
+dev.detekt.report.xml.XmlOutputReport
